fixed automatic log in problem
