+++ dapps/user-dashboard.mdx
@@ -3,6 +3,6 @@ sidebar_position: 0
 title: User Dashboard
 ---
 
-import Login from "./user-dashboard"
+import Login from "../src/components/molecules/user-dashboard"
 
 <Login />
