+++ package.json
@@ -19,11 +19,11 @@
     "@docusaurus/core": "^2.2.0",
     "@docusaurus/preset-classic": "^2.2.0",
     "@mdx-js/react": "^1.6.22",
-    "@soroban-react/chains": "^5.0.3",
-    "@soroban-react/core": "^5.0.3",
-    "@soroban-react/events": "^5.0.3",
-    "@soroban-react/freighter": "^5.0.3",
-    "@soroban-react/types": "^5.0.3",
+    "@soroban-react/chains": "^5.3.12",
+    "@soroban-react/core": "^5.3.12",
+    "@soroban-react/events": "^5.3.12",
+    "@soroban-react/freighter": "^5.3.12",
+    "@soroban-react/types": "^5.3.12",
     "clsx": "^1.2.1",
     "docusaurus-plugin-sass": "^0.2.2",
     "mdx-mermaid": "^1.3.2",
