+++ src/components/molecules/user-dashboard/style.module.css
@@ -1,0 +1,58 @@
+.card {
+  /* border: 1px solid black; */
+  width: 700px;
+  padding: 20px;
+  border-radius: 15px; /* soft rounded corners */
+  box-shadow: 0 2px 8px 0 rgba(0, 0, 0, 0.2);
+  transition: 0.3s;
+  background: white;
+  margin-top: 20px;
+}
+
+.container {
+  display: flex;
+  justify-content: flex-start;
+  gap: 20px; /* Optional: Adds some space between the cards */
+}
+
+.card h3 {
+  color: #333;
+  font-size: 24px;
+}
+
+.card p,
+.card h4,
+.card ul {
+  color: #666;
+}
+
+.card ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.card ul li {
+  padding: 5px 0;
+}
+
+.card:hover {
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+}
+
+.button {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding: 16px 12px;
+  gap: 8px;
+  background: #1a1523;
+  border-radius: 8px;
+  border: 0;
+  height: 38px;
+  font-weight: 600;
+  font-size: 14px;
+  line-height: 22px;
+  color: #ffffff;
+  cursor: pointer;
+}
